pyproject.toml
[tool.poetry]
name = "sudokusolver"
version = "0.1.0"
description = ""
authors = ["raphaellndr <raph.landure@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
django = "5.1.6"
djangorestframework = "3.15.2"
drf-spectacular = "0.28.0"
django-cors-headers = "4.7.0"
djangorestframework-simplejwt = "5.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
mypy = "^1.15.0"
pre-commit = "^4.1.0"
pylint = "^3.3.4"
django-stubs = "^5.1.3"
djangorestframework-stubs = "^3.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
